Add vector table
